+++ Konsom.Application/CommandAndQuery/Tags/Commands/DeleteTag/DeleteTagCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using FluentValidation.Results;
 using Konsom.Application.Interfaces;
 using MediatR;
 
@@ -17,8 +18,21 @@ namespace Konsom.Application.CommandAndQuery.Tags.Commands.DeleteTag
 
         public async Task<Unit> Handle(DeleteTagCommand request, CancellationToken cancellationToken)
         {
-            await _repository.Delete(request.Id);
-            return Unit.Value;
+            try
+            {
+                var deleteTagCommandValidator = new DeleteTagCommandValidator();
+                ValidationResult result = deleteTagCommandValidator.Validate(request);
+                if (!result.IsValid)
+                {
+                    throw new Exception("Данные не валидны");
+                }
+                await _repository.Delete(request.Id);
+                return Unit.Value;
+            }
+            catch (Exception)
+            {
+                throw;
+            }
         }
     }
 }
